+++ newspack-event-aggregator/newspack-event-aggregator.php
@@ -2,7 +2,7 @@
 /**
  * Plugin Name: Newspack Event Aggregator
  * Description: Aggregates logs from remote Event Logger servers via SSE.
- * Version: 2.4.17
+ * Version: 2.4.18
  * Author: Automattic
  * Author URI: https://newspack.com/
  * License: GPL-2.0-or-later
@@ -34,7 +34,7 @@ if ( ! \class_exists( 'Newspack_Event_Logger\\Firehose' ) ) {
 }
 
 if ( ! \defined( 'NEWSPACK_EVENT_AGGREGATOR_VERSION' ) ) {
-	\define( 'NEWSPACK_EVENT_AGGREGATOR_VERSION', '2.4.17' );
+	\define( 'NEWSPACK_EVENT_AGGREGATOR_VERSION', '2.4.18' );
 }
 
 // Define the plugin directory and URL.